# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: PermsCraft REST API
  description: |
    Embedded REST API for the PermsCraft Minecraft permission plugin.

    ## Authentication
    Every request must include a valid API key in the `Authorization` header:
    ```
    Authorization: Bearer <api-key>
    ```
    Keys are managed in-game or from console:
    ```
    /pc apikey create <label> [scopes...]
    /pc apikey list
    /pc apikey revoke <label>
    ```

    ## Scopes
    | Scope    | Description |
    |----------|-------------|
    | `read`   | GET endpoints for groups, users, server info |
    | `write`  | POST/PUT/DELETE for groups and users |
    | `log`    | Access to audit log endpoints |
    | `sync`   | Trigger cache flush and sync operations |
    | `backup` | Export, import, and restore backups |
    | `admin`  | All scopes combined |

    ## Security
    > ⚠️ The API transmits keys over plain HTTP by default.
    > It is strongly recommended to use a TLS reverse proxy (nginx/Caddy)
    > and bind to `127.0.0.1` in `config.yml`.
    > See the [TLS setup guide](https://github.com/your-repo/wiki/REST-API-TLS).

    ## Rate Limiting
    Default: 120 requests per minute per IP. Configurable in `config.yml`.
  version: 1.0.0
  contact:
    name: PermsCraft
    url: https://github.com/your-repo
servers:
  - url: http://localhost:4567/api/v1
    description: Local server (default)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Group 'admin' not found."
    Group:
      type: object
      properties:
        name:
          type: string
          example: admin
        weight:
          type: integer
          example: 100
        prefix:
          type: string
          example: "&c[Admin] "
        suffix:
          type: string
          example: ""
        parents:
          type: array
          items:
            type: string
          example: ["default", "vip"]
        permissions:
          type: array
          items:
            type: string
          example: ["essentials.fly", "-essentials.ban"]
        meta:
          type: object
          additionalProperties:
            type: string
...
paths:
...
  /groups:
    get:
      summary: List all groups
      tags: [Groups]
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  count:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a group
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: vip
                weight:
                  type: integer
                  default: 0
                  example: 10
      responses:
        "201":
          description: Group created
        "400":
          description: Invalid input or group already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /backup/import:
    post:
      summary: Import a backup (overwrites current data)
      tags: [Backup]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Import complete
        "400":
          description: Invalid backup format
...
  /backup/restore/{name}:
    post:
      summary: Restore a named local snapshot
      tags: [Backup]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restore complete
        "404":
          description: Snapshot not found

  # ── Server ──────────────────────────────────────────────────────────────────
...